modify: move course from common/ to common/components/course/
